Renames _multiprocessing.py to not conflict with standard library name. Fixes #127
